Test Equipment
Test everything related to equipment and containers except reassign, because that uses powersync which is tricky

+++ frontend/src/lib/services/equipment.ts
@@ -4,7 +4,7 @@ import { EquipmentRecord, ContainerRecord } from "../../types/db";
 import { Equipment, Container } from "../../types/models";
 import { AbstractPowerSyncDatabase } from "@powersync/react-native";
 import { Queries } from "../powersync/queries";
-import { handleSupabaseError } from "./util";
+import { handleSupabaseError, isManager } from "./util";
 
 export interface IEquipmentService {
   deleteEquipment(equipment: Equipment): Promise<void>;
@@ -32,12 +32,20 @@ export interface IEquipmentService {
 
 export class EquipmentService implements IEquipmentService {
   async deleteEquipment(equipment: Equipment): Promise<void> {
+    if (!(await isManager(equipment.organizationId))) {
+      throw new Error("Only managers can delete equipment.");
+    }
+
     const ids = equipment.records.map((r) => r.id);
     const { error } = await supabase.from("equipment").delete().in("id", ids);
     if (error) handleSupabaseError(error);
   }
 
   async deleteContainer(container: Container): Promise<void> {
+    if (!(await isManager(container.organizationId))) {
+      throw new Error("Only managers can delete containers.");
+    }
+
     const { error: conError } = await supabase
       .from("containers")
       .delete()
@@ -49,6 +57,10 @@ export class EquipmentService implements IEquipmentService {
     quantity: number,
     data: Omit<EquipmentRecord, "id" | "last_updated_date">,
   ): Promise<void> {
+    if (!(await isManager(data.organization_id))) {
+      throw new Error("Only managers can create equipment.");
+    }
+
     const timestamp = new Date().toISOString();
     const items = Array.from({ length: quantity }).map(() => ({
       id: generateUUID(),
@@ -64,6 +76,10 @@ export class EquipmentService implements IEquipmentService {
     quantity: number,
     data: Omit<ContainerRecord, "id" | "last_updated_date">,
   ): Promise<void> {
+    if (!(await isManager(data.organization_id))) {
+      throw new Error("Only managers can create containers.");
+    }
+
     const timestamp = new Date().toISOString();
     const items = Array.from({ length: quantity }).map(() => ({
       id: generateUUID(),
